Fix content/setContent; add access to editor
Currently this.find('.md-editor') will not work since this refers to the textarea which does not contain .md-editor (it comes before).

This change makes access to editor more explicit by putting it in the data for this object. It also adds the method 'editor' which provides direct access to the ace editor object.